Added whitespace between comma and <milestone>.
It is not obvious how a <milestone> should be interpreted with respect to whitespace
(i.e. should it 'count' as whitespace between text or not), but since whitespace is
added liberally before a milestone, after a milestone and sometimes both before and
after a milestone in most of the texts in the collection, it makes no sense for
whitespace to be missing both before after milestones that immediately follow a
comma in Caes. Gal. This inserts a single space before the milestone.